separated out structure factors
