+++ src/sas/sascalc/shape2sas/Shape2SAS.py
@@ -6,8 +6,9 @@ import numpy as np
 from typing import Optional, List
 from dataclasses import dataclass, field
 
+from sas.sascalc.shape2sas.StructureFactor import StructureFactor
 from sas.sascalc.shape2sas.helpfunctions import (
-    GenerateAllPoints, WeightedPairDistribution, StructureFactor, ITheoretical, IExperimental, Qsampling,
+    GenerateAllPoints, WeightedPairDistribution, ITheoretical, IExperimental, Qsampling,
     plot_2D, plot_results, generate_pdb
 )
 
@@ -87,9 +88,6 @@ class TheoreticalScattering:
     I0: np.ndarray
     I: np.ndarray
     S_eff: np.ndarray
-    r: np.ndarray #pair distance distribution
-    pr: np.ndarray #pair distance distribution
-    pr_norm: np.ndarray #normalized pair distance distribution
 
 
 @dataclass
@@ -145,7 +143,7 @@ def getTheoreticalScattering(scalc: TheoreticalScatteringCalculation) -> Theoret
 
     I = I_theory.calc_Iq(Pq, S_eff, sys.sigma_r)
 
-    return TheoreticalScattering(q=q, I=I, I0=I0, S_eff=S_eff, r=r, pr=pr, pr_norm=pr_norm)
+    return TheoreticalScattering(q=q, I=I, I0=I0, S_eff=S_eff)
 
 
 def getSimulatedScattering(scalc: SimulateScattering) -> SimulatedScattering:
