+++ src/sas/sascalc/shape2sas/helpfunctions.py
@@ -2,15 +2,9 @@ import numpy as np
 import matplotlib.pyplot as plt
 from sas.sascalc.shape2sas.Typing import *
 from sas.sascalc.shape2sas.models import *
+from sas.sascalc.shape2sas.Math import sinc
 
 ################################ Shape2SAS helper functions ###################################
-def sinc(x) -> np.ndarray:
-    """
-    function for calculating sinc = sin(x)/x
-    numpy.sinc is defined as sinc(x) = sin(pi*x)/(pi*x)
-    """
-    return np.sinc(x / np.pi)   
-
 class Qsampling:
     def onQsampling(qmin: float, qmax: float, Nq: int) -> np.ndarray:
         """Returns uniform q sampling"""
@@ -581,296 +575,6 @@ class WeightedPairDistribution:
             f.write('# %-17s %-17s\n' % ('r','p(r)'))
             for i in range(Nbins):
                 f.write('  %-17.5e %-17.5e\n' % (r[i], pr_norm[i]))
-
-
-class StructureDecouplingApprox:
-    def __init__(self, q: np.ndarray, 
-                 x_new: np.ndarray, 
-                 y_new: np.ndarray, 
-                 z_new: np.ndarray, 
-                 p_new: np.ndarray):
-        self.q = q
-        self.x_new = x_new
-        self.y_new = y_new
-        self.z_new = z_new
-        self.p_new = p_new
-
-    def calc_com_dist(self) -> np.ndarray:
-        """ 
-        calc contrast-weighted com distance
-        """
-        w = np.abs(self.p_new)
-
-        if np.sum(w) == 0:
-            w = np.ones(len(self.x_new))
-
-        x_com, y_com, z_com = np.average(self.x_new, weights=w), np.average(self.y_new, weights=w), np.average(self.z_new, weights=w)
-        dx, dy, dz = self.x_new - x_com, self.y_new - y_com, self.z_new - z_com
-        com_dist = np.sqrt(dx**2 + dy**2 + dz**2)
-
-        return com_dist
-
-    def calc_A00(self) -> np.ndarray:
-        """
-        calc zeroth order sph harm, for decoupling approximation
-        """
-        d_new = self.calc_com_dist()
-        M = len(self.q)
-        A00 = np.zeros(M)
-        
-        for i in range(M):
-            qr = self.q[i] * d_new
-
-            A00[i] = sum(self.p_new * sinc(qr))
-        A00 = A00 / A00[0] # normalise, A00[0] = 1
-
-        return A00
-    
-    def decoupling_approx(self, Pq: np.ndarray, S: np.ndarray) -> np.ndarray:
-        """
-        modify structure factor with the decoupling approximation
-        for combining structure factors with non-spherical (or polydisperse) particles
-
-        see, for example, Larsen et al 2020: https://doi.org/10.1107/S1600576720006500
-        and refs therein
-
-        input
-        q
-        x,y,z,p    : coordinates and contrasts
-        Pq         : form factor
-        S          : structure factor
-
-        output
-        S_eff      : effective structure factor, after applying decoupl. approx
-        """
-
-        A00 = self.calc_A00()
-        const = 1e-3 # add constant in nominator and denominator, for stability (numerical errors for small values dampened)
-        Beta = (A00**2 + const) / (Pq + const)
-        S_eff = 1 + Beta * (S - 1)
-
-        return S_eff
-
-
-'''#template for the structure factor classes
-class <NAME OF STRUCTURE FACTOR HERE>(StructureDecouplingApprox):
-    def __init__(self, q: np.ndarray,
-                    x_new: np.ndarray,
-                    y_new: np.ndarray,
-                    z_new: np.ndarray,
-                    p_new: np.ndarray,
-                    par: List[float]):
-            super(<NAME OF STRUCTURE FACTOR HERE>, self).__init__(q, x_new, y_new, z_new, p_new)
-            self.q = q
-            self.x_new = x_new
-            self.y_new = y_new
-            self.z_new = z_new
-            self.p_new = p_new
-            self.par = par[0] <WRITE YOUR PARAMETERS HERE>
-
-    def structure_eff(self, Pq: np.ndarray) -> np.ndarray:
-        S = <STRUCTURE FACTOR HERE>
-        S_eff = self.decoupling_approx(Pq, S)
-
-        return S_eff
-'''
-
-
-class HardSphereStructure(StructureDecouplingApprox):
-    def __init__(self, q: np.ndarray, 
-                 x_new: np.ndarray, 
-                 y_new: np.ndarray, 
-                 z_new: np.ndarray, 
-                 p_new: np.ndarray, 
-                 par: List[float]):
-        super(HardSphereStructure, self).__init__(q, x_new, y_new, z_new, p_new)
-        self.q = q
-        self.x_new = x_new
-        self.y_new = y_new
-        self.z_new = z_new
-        self.p_new = p_new
-        self.conc = par[0]
-        self.R_HS = par[1]
-
-    def calc_S_HS(self) -> np.ndarray:
-        """
-        calculate the hard-sphere structure factor
-        calls function calc_G()
-
-        input
-        q       : momentum transfer
-        eta     : volume fraction
-        R       : estimation of the hard-sphere radius
-
-        output
-        S_HS    : hard-sphere structure factor
-        """
-
-        if self.conc > 0.0:
-            A = 2 * self.R_HS * self.q 
-            G = self.calc_G(A, self.conc)
-            S_HS = 1 / (1 + 24 * self.conc * G / A) #percus-yevick approximation for
-        else:                         #calculating the structure factor
-            S_HS = np.ones(len(self.q))
-
-        return S_HS
-    
-    @staticmethod
-    def calc_G(A: np.ndarray, eta: float) -> np.ndarray:
-        """ 
-        calculate G in the hard-sphere potential
-
-        input
-        A  : 2*R*q
-        q  : momentum transfer
-        R  : hard-sphere radius
-        eta: volume fraction
-
-        output:
-        G  
-        """
-
-        a = (1 + 2 * eta)**2 / (1 - eta)**4
-        b = -6 * eta * (1 + eta / 2)**2/(1 - eta)**4 
-        c = eta * a / 2
-        sinA = np.sin(A)
-        cosA = np.cos(A)
-        fa = sinA - A * cosA
-        fb = 2 * A * sinA + (2 - A**2) * cosA-2
-        fc = -A**4 * cosA + 4 * ((3 * A**2 - 6) * cosA + (A**3 - 6 * A) * sinA + 6)
-        G = a * fa / A**2 + b * fb / A**3 + c * fc / A**5
-
-        return G
-
-    def structure_eff(self, Pq: np.ndarray) -> np.ndarray:
-        S = self.calc_S_HS()
-        S_eff = self.decoupling_approx(Pq, S)
-        return S_eff 
-
-
-class Aggregation(StructureDecouplingApprox):
-    def __init__(self, q: np.ndarray, 
-                 x_new: np.ndarray, 
-                 y_new: np.ndarray, 
-                 z_new: np.ndarray, 
-                 p_new: np.ndarray, 
-                 par: List[float]):
-        super(Aggregation, self).__init__(q, x_new, y_new, z_new, p_new)
-        self.q = q
-        self.x_new = x_new
-        self.y_new = y_new
-        self.z_new = z_new
-        self.p_new = p_new
-        self.Reff = par[0]
-        self.Naggr = par[1]
-        self.fracs_aggr = par[2]
-
-    def calc_S_aggr(self) -> np.ndarray:
-        """
-        calculates fractal aggregate structure factor with dimensionality 2
-
-        S_{2,D=2} in Larsen et al 2020, https://doi.org/10.1107/S1600576720006500
-
-        input 
-        q      :
-        Naggr  : number of particles per aggregate
-        Reff   : effective radius of one particle 
-
-        output
-        S_aggr :
-        """
-
-        qR = self.q * self.Reff
-        S_aggr = 1 + (self.Naggr - 1)/(1 + qR**2 * self.Naggr / 3)
-
-        return S_aggr
-
-    def structure_eff(self, Pq: np.ndarray) -> np.ndarray:
-        """Return effective structure factor for aggregation"""
-
-        S = self.calc_S_aggr()
-        S_eff = self.decoupling_approx(Pq, S)
-        S_eff = (1 - self.fracs_aggr) + self.fracs_aggr * S_eff
-        return S_eff
-
-
-class NoStructure(StructureDecouplingApprox):
-    def __init__(self, q: np.ndarray, 
-                 x_new: np.ndarray, 
-                 y_new: np.ndarray, 
-                 z_new: np.ndarray, 
-                 p_new: np.ndarray, 
-                 par: Any):
-        super(NoStructure, self).__init__(q, x_new, y_new, z_new, p_new)
-        self.q = q
-
-    def structure_eff(self, Pq: Any) -> np.ndarray:
-        """Return effective structure factor for no structure"""
-        return np.ones(len(self.q))
-
-
-class StructureFactor:
-    def __init__(self, q: np.ndarray, 
-                 x_new: np.ndarray, 
-                 y_new: np.ndarray, 
-                 z_new: np.ndarray, 
-                 p_new: np.ndarray,
-                 Stype: str,
-                 par: Optional[List[float]]):
-        self.q = q
-        self.x_new = x_new
-        self.y_new = y_new
-        self.z_new = z_new
-        self.p_new = p_new
-        self.Stype = Stype
-        self.par = par
-        self.setAvailableStructureFactors()
-
-    def setAvailableStructureFactors(self):
-        """Available structure factors"""
-        self.structureFactor = {
-            'HS': HardSphereStructure,
-            'Hard Sphere': HardSphereStructure,
-            'aggregation': Aggregation,
-            'Aggregation': Aggregation,
-            'None': NoStructure
-        }
-    
-    def getStructureFactorClass(self):
-        """Return chosen structure factor"""
-        if self.Stype in self.structureFactor:
-            return self.structureFactor[self.Stype](self.q, self.x_new, self.y_new, self.z_new, self.p_new, self.par)
-        
-        else:
-            try:
-                return globals()[self.Stype](self.q, self.x_new, self.y_new, self.z_new, self.p_new, self.par)
-            except KeyError:
-                ValueError(f"Structure factor '{self.Stype}' was not found in structureFactor or global scope.")
-
-    @staticmethod
-    def getparname(name: str) -> List[str]:
-        """Return the name of the parameters"""
-        pars = {
-            'HS': {'conc': 0.02,'r_hs': 50},
-            'Hard Sphere': {'conc': 0.02,'r_hs': 50},
-            'Aggregation': {'R_eff': 50, 'N_aggr': 80, 'frac': 0.1},
-            'aggregation': {'R_eff': 50, 'N_aggr': 80, 'frac': 0.1},
-            'None': {}
-        }
-        return pars[name]
-
-    @staticmethod
-    def save_S(q: np.ndarray, S_eff: np.ndarray, Model: str):
-        """ 
-        save S to file
-        """
-
-        with open('Sq%s.dat' % Model,'w') as f:
-            f.write('# Structure factor, S(q), used in: I(q) = P(q)*S(q)\n')
-            f.write('# Default: S(q) = 1.0\n')
-            f.write('# %-17s %-17s\n' % ('q','S(q)'))
-            for (q_i, S_i) in zip(q, S_eff):
-                f.write('  %-17.5e%-17.5e\n' % (q_i, S_i))
 
 
 class ITheoretical:
