+++ src/sas/sascalc/shape2sas/structure_factors/__init__.py
@@ -1,0 +1,10 @@
+from .Aggregation import Aggregation
+from .HardSphereStructure import HardSphereStructure
+from .NoStructure import NoStructure
+from .StructureDecouplingApprox import StructureDecouplingApprox
+
+__all__ = [
+    'Aggregation',
+    'HardSphereStructure',
+    'NoStructure'
+]
